add authorization module

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,14 +1,17 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantCommandHandler(
         ILogger<DeleteRestaurantCommand> logger,
-        IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+        IRestaurantsRepository restaurantsRepository,
+        IRestaurantAuthorizationService restaurnatAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -18,6 +21,10 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
 
             if (restaurant == null)
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+            if (!restaurnatAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+                throw new ForbidException();
+            
 
             await restaurantsRepository.Delete(restaurant, cancellationToken);
         }
